requirements.txt
requests
python-firebase